BasicFlaskAuth/requirements.txt
astroid==2.2.5
Click==7.0
ecdsa==0.13.2
Flask==1.0.2
future==0.17.1
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.0
MarkupSafe==1.1.1
mccabe==0.6.1
pycryptodome==3.6.6
pylint==2.3.1
python-jose-cryptodome==1.3.2
six==1.12.0
typed-ast==1.3.5
Werkzeug==0.15.2
wrapt==1.11.1
Flask-Cors==3.0.8
